feat: remove approval for restaurant adjust sql init

+++ PayoEat-BE/src/main/java/com/example/PayoEat_BE/service/user/UserService.java
@@ -3,8 +3,10 @@ package com.example.PayoEat_BE.service.user;
 
 import com.example.PayoEat_BE.exceptions.InvalidException;
 import com.example.PayoEat_BE.exceptions.NotFoundException;
+import com.example.PayoEat_BE.model.Restaurant;
 import com.example.PayoEat_BE.model.User;
 import com.example.PayoEat_BE.model.VerificationToken;
+import com.example.PayoEat_BE.repository.RestaurantRepository;
 import com.example.PayoEat_BE.repository.UserRepository;
 import com.example.PayoEat_BE.repository.VerificationTokenRepository;
 import com.example.PayoEat_BE.request.menu.CreateUserRequest;
@@ -31,9 +33,9 @@ import java.util.regex.Pattern;
 public class UserService implements IUserService{
     private final UserRepository userRepository;
     private final VerificationTokenRepository tokenRepository;
-    private final ModelMapper modelMapper;
     private final PasswordEncoder passwordEncoder;
     private final EmailService emailService;
+    private final RestaurantRepository restaurantRepository;
 
     private static final String EMAIL_REGEX = "^[A-Za-z0-9+_.-]+@(.+)$";
     private static final Pattern EMAIL_PATTERN = Pattern.compile(EMAIL_REGEX);
@@ -98,10 +100,12 @@ public class UserService implements IUserService{
 
         user.setIsActive(true);
 
-        userRepository.activateUser(user.getId());
-        tokenRepository.delete(optionalToken.getId());
+        if (restaurantRepository.setRestaurantToActive(user.getId()) > 0) {
+            userRepository.activateUser(user.getId());
+            tokenRepository.delete(optionalToken.getId());
+        }
 
-        return "User confirmed successfully.";
+        return "User confirmed successfully, the restaurant is now active";
     }
 
     @Override
